requirements/base.txt
-r base.txt
coverage
tox
model_mommy
django-nose
